login: move category topics and link preview into their own services
/category-topics/ moves to services/topics.get_category_topics -- redis
get-or-set plus a query, with the standalone session and 1-day TTL kept
verbatim.

/link-preview/ moves to services/link_preview.get_link_preview, next to the
request_text fetch it already used. The host allowlist and the OpenGraph scrape
belong with the fetch, not in api/.

The link-preview endpoint keeps its unused `ctx` parameter on purpose, with a
comment: get_request_context is what puts the OAuth2PasswordBearer security
entry on the operation, so dropping it would change the OpenAPI document.
